fixed a bug where export would fail because file is already in use
this happened when using the open video functionality via the menu item or by dragging a video on the dock. If the same video was used twice the second time the export would fail with the error message written above. This happened because a file with that file name already existed due to the first video export. A timestamp is now added to the file name to resolve this issue.